requirements.txt
requests ~= 2.31.0
chromadb ~= 0.5.0
openai ~= 1.33.0
pathlib ~=  1.0.1
langchain ~= 0.2.3
python-docx ~= 1.1.2